% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constant.R
\docType{data}
\name{DISEASE_CATEGORIES}
\alias{DISEASE_CATEGORIES}
\title{Define disease categories based on ICD-10 code ranges}
\format{
An object of class \code{list} of length 22.
}
\usage{
DISEASE_CATEGORIES
}
\description{
Define disease categories based on ICD-10 code ranges
}
\keyword{datasets}
