% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R, R/prevalence_functions.R
\name{get_category_count}
\alias{get_category_count}
\title{Get patient counts for a disease category}
\usage{
get_category_count(conn, cdm_schema, category)

get_category_count(conn, cdm_schema, category)
}
\arguments{
\item{conn}{Database connection}

\item{cdm_schema}{CDM schema name}

\item{category}{Disease category name}
}
\value{
Number of unique patients

Number of unique patients
}
\description{
Get patient counts for a disease category

Get patient counts for a disease category
}
